refactor(cleanup): remove kernel-adoption orphans; thread PlatformSpec through Result
Verification follow-ups (user-directed):

- pkg/module trimmed to the metadata types + CanonicalModuleRef: the
  Instance and Module structs (and MatchComponents) had no callers after
  kernel adoption; pkg/loader's DetectInstanceKind likewise
  (resolveInstanceFile helper relocated into instance_file.go)
- seven removal/partial delta specs retire the main-spec surface the
  deletions orphan (module-instance-type, module-raw-field,
  module-instance-parsing, module-instance-processing, loader-api,
  instance-file-loading, pkg-types)
- reviewer suggestions implemented: config.LoadPlatformFile returns the
  compiled value so DecodeFile reads/compiles once; the resolved
  synth.PlatformInput threads through render.Result (PlatformSpec) and
  the D12 seeding hook uses it directly — no file re-read, no TOCTOU
